migrate onChange logic to the new layer
