+++ packages/client/src/components/advanced/Dropdowns/AttributeMultiDropdown.tsx
@@ -2,6 +2,7 @@ import { allEntities } from "@shared/dictionaries/entity";
 import { EntityEnums } from "@shared/enums";
 import { BaseDropdown } from "components";
 import React from "react";
+import { DropdownItem } from "types";
 
 interface AttributeMultiDropdown<T = string> {
   width?: number | "full";
@@ -27,6 +28,8 @@ export const AttributeMultiDropdown = <T extends string>({
 
   loggerId,
 }: AttributeMultiDropdown<T>) => {
+  const getValues = (items: DropdownItem[]) => items.map((i) => i.value as T);
+
   return (
     <BaseDropdown
       // TODO: hopefully get rid of this prop
@@ -40,7 +43,37 @@ export const AttributeMultiDropdown = <T extends string>({
       value={[allEntities]
         .concat(options)
         .filter((o) => value.includes(o.value as T))}
-      onChange={(items) => onChange(items.map((i) => i.value as T))}
+      onChange={(selectedOptions, event) => {
+        // kdyz je neco vybrany = aspon jeden option
+        if (selectedOptions !== null && selectedOptions.length > 0) {
+          if (event?.action === "remove-value") {
+            return onChange([]);
+          }
+          if (
+            selectedOptions[selectedOptions.length - 1].value ===
+            allEntities.value
+          ) {
+            // kdyz vyberu all option
+            return onChange(getValues(options));
+          }
+          let result: DropdownItem[] = [];
+          // jsou vybrany vsechny
+          if (selectedOptions.length === options.length) {
+            // kdyz jsou vybrany vsechny
+            if (selectedOptions.includes(allEntities)) {
+              //
+              result = selectedOptions.filter(
+                (option: { label: string; value: string }) =>
+                  option.value !== allEntities.value
+              );
+            } else if (event?.action === "select-option") {
+              result = options;
+            }
+            return onChange(getValues(result));
+          }
+        }
+        return onChange(getValues(selectedOptions));
+      }}
       disabled={disabled}
       loggerId={loggerId}
     />
