+++ packages/client/src/components/advanced/Dropdowns/EntityMultiDropdown.tsx
@@ -1,6 +1,7 @@
 import { allEntities, empty } from "@shared/dictionaries/entity";
 import { BaseDropdown } from "components";
 import React from "react";
+import { DropdownItem } from "types";
 
 interface EntityMultiDropdown<T = string> {
   width?: number | "full";
@@ -24,6 +25,8 @@ export const EntityMultiDropdown = <T extends string>({
 
   loggerId,
 }: EntityMultiDropdown<T>) => {
+  const getValues = (items: DropdownItem[]) => items.map((i) => i.value as T);
+
   return (
     <BaseDropdown
       entityDropdown
@@ -33,7 +36,34 @@ export const EntityMultiDropdown = <T extends string>({
       value={[empty, allEntities]
         .concat(options)
         .filter((o) => value.includes(o.value as T))}
-      onChange={(items) => onChange(items.map((i) => i.value as T))}
+      onChange={(selectedOptions, event) => {
+        // kdyz je neco vybrany = aspon jeden option
+        if (selectedOptions !== null && selectedOptions.length > 0) {
+          if (
+            selectedOptions[selectedOptions.length - 1].value ===
+            allEntities.value
+          ) {
+            // kdyz vyberu all option
+            return onChange(getValues(options));
+          }
+          let result: DropdownItem[] = [];
+          // jsou vybrany vsechny
+          if (selectedOptions.length === options.length) {
+            // kdyz jsou vybrany vsechny
+            if (selectedOptions.includes(allEntities)) {
+              //
+              result = selectedOptions.filter(
+                (option: { label: string; value: string }) =>
+                  option.value !== allEntities.value
+              );
+            } else if (event?.action === "select-option") {
+              result = options;
+            }
+            return onChange(getValues(result));
+          }
+        }
+        return onChange(getValues(selectedOptions));
+      }}
       placeholder={placeholder}
       noOptionsMessage={noOptionsMessage}
       disabled={disabled}
